+++ samples/JwtSample/Properties/launchSettings.json
@@ -3,7 +3,7 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:57780/",
+      "applicationUrl": "http://localhost:8719/",
       "sslPort": 0
     }
   },
