+++ samples/SocketsSample/Properties/launchSettings.json
@@ -3,7 +3,7 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:59847/",
+      "applicationUrl": "http://localhost:8718/",
       "sslPort": 0
     }
   },
